requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
pydantic-settings==2.10.1
sqlalchemy[asyncio]==2.0.43
asyncpg==0.30.0
alembic==1.16.5
redis[hiredis]==6.4.0
pymilvus==2.6.1
minio==7.2.18
httpx==0.28.1
langgraph==0.6.7
python-multipart==0.0.20
pytest==8.4.2
pytest-asyncio==1.2.0
